Add test case for GMB client creation
