+++ .gitignore
@@ -11,3 +11,4 @@ phpunit.xml
 psalm.xml
 vendor
 .env
+.env.test
